+++ src/chainType.ts
@@ -1,0 +1,6 @@
+export enum ChainType {
+  LLM_CHAIN = "llm_chain",
+  VAULT_QA_CHAIN = "vault_qa",
+  COPILOT_PLUS_CHAIN = "copilot_plus",
+  PROJECT_CHAIN = "project",
+}
